+++ src/Camera/cameragc.cpp
@@ -259,7 +259,7 @@ void CPolarSubCamera::loadAfter()
 	if ((unk64 & 0x1000) && gpMarDirector->unk7D == 1) {
 		gpMarDirector->fireStartDemoCamera(
 		    cJetCoasterDemoBckName, nullptr, -1, 0.0f, true,
-		    &JetCoasterDemoCallBack, (u32)this, nullptr, JDrama::TFlagT<u16>());
+		    &JetCoasterDemoCallBack, (u32)(uintptr_t)this, nullptr, JDrama::TFlagT<u16>());
 	} else {
 		if (!JKRGetResource(cStartCamBckFileName))
 			calcInHouseNo_(true);
