+++ src/JSystem/J3D/J3DGraphLoader/J3DMaterialFactory.cpp
@@ -14,7 +14,7 @@ J3DMaterialFactory::J3DMaterialFactory(const J3DMaterialBlock& block)
 	mpMaterialID = JSUConvertOffsetToPtr<u16>(&block, block.mpMaterialID);
 
 	if (block.mpIndInitData != nullptr
-	    && ((u32)block.mpIndInitData - (u32)block.mpNameTable) > 4)
+	    && ((u32)(uintptr_t)block.mpIndInitData - (u32)(uintptr_t)block.mpNameTable) > 4)
 		mpIndInitData = JSUConvertOffsetToPtr<J3DIndInitData>(
 		    &block, block.mpIndInitData);
 	else
