+++ src/JSystem/J3D/J3DGraphLoader/J3DShapeFactory.cpp
@@ -6,17 +6,17 @@
 J3DShapeFactory::J3DShapeFactory(const J3DShapeBlock& block)
 {
 	mpShapeInitData = JSUConvertOffsetToPtr<J3DShapeInitData>(
-	    &block, (u32)block.mpShapeInitData);
-	mpIndexTable  = JSUConvertOffsetToPtr<u16>(&block, (u32)block.mpIndexTable);
+	    &block, (u32)(uintptr_t)block.mpShapeInitData);
+	mpIndexTable  = JSUConvertOffsetToPtr<u16>(&block, (u32)(uintptr_t)block.mpIndexTable);
 	mpVtxDescList = JSUConvertOffsetToPtr<GXVtxDescList>(
-	    &block, (u32)block.mpVtxDescList);
-	mpMtxTable = JSUConvertOffsetToPtr<u16>(&block, (u32)block.mpMtxTable);
+	    &block, (u32)(uintptr_t)block.mpVtxDescList);
+	mpMtxTable = JSUConvertOffsetToPtr<u16>(&block, (u32)(uintptr_t)block.mpMtxTable);
 	mpDisplayListData
-	    = JSUConvertOffsetToPtr<u8>(&block, (u32)block.mpDisplayListData);
+	    = JSUConvertOffsetToPtr<u8>(&block, (u32)(uintptr_t)block.mpDisplayListData);
 	mpMtxInitData = JSUConvertOffsetToPtr<J3DShapeMtxInitData>(
-	    &block, (u32)block.mpMtxInitData);
+	    &block, (u32)(uintptr_t)block.mpMtxInitData);
 	mpDrawInitData = JSUConvertOffsetToPtr<J3DShapeDrawInitData>(
-	    &block, (u32)block.mpDrawInitData);
+	    &block, (u32)(uintptr_t)block.mpDrawInitData);
 	mpVcdVatCmdBuffer = nullptr;
 }
