+++ src/JSystem/JAudio/JASystem/JASHeapCtrl.cpp
@@ -329,9 +329,9 @@ namespace Kernel {
 			mEnd  = 0;
 			unk10 = 0;
 		} else {
-			mStart = (u8*)ALIGN_NEXT((u32)mem_start, 0x20);
+			mStart = (u8*)ALIGN_NEXT((uintptr_t)mem_start, 0x20);
 			mEnd   = mStart;
-			mSize  = mem_size - ((u32)mem_start & 0x1F);
+			mSize  = mem_size - ((uintptr_t)mem_start & 0x1F);
 			unk10  = 0;
 		}
 	}
